% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{nowarning}
\alias{nowarning}
\title{Dummy function that does NOT cause a warning}
\usage{
nowarning()
}
\description{
Dummy function whose sole purpose is to reproduce the OMP warning.  This
one is ok.
from data.table.
}
