% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{yeswarning}
\alias{yeswarning}
\title{Dummy function that DOES cause a warning}
\usage{
yeswarning()
}
\description{
Dummy function whose sole purpose is to reproduce the OMP warning.
This one DOES cause a warning.
from data.table.
}
